[v 1.0.0 ] [fix: 리뷰반영 - 폴더 및 패키지 구조 변경   ]
