+++ iTime/Domain/Sources/Repository/NetworkRepository/AuthenticationRepository/Entities/UserSession.swift
