+++ iTime/Domain/Sources/Repository/NetworkRepository/AuthenticationRepository/Result/ThirdPartyAuthenticationResult.swift
