+++ iTime/Feature/Sources/EditorImpl/BookmarkEditor/BookmarkEditor/BookmarkCollectionDelegateDataSource/SavedItemSectionDelegateDataSource.swift
@@ -6,7 +6,7 @@
 //
 
 import SharedUI
-import BookmarkUsecase
+import Usecase
 
 // MARK: - SavedItemSectionListener
